intoduce SessionRewardService.cs

+++ src/RewardFlow.API/Rewards/Data/Subject.cs
@@ -6,7 +6,7 @@ public class Subject
     public string Name { get; set; } = null!;
     public bool IsTheoretical { get; set; }
     public bool IsPractical { get; set; }
-    public float SubjectPrice { get; set; }
+    public decimal SubjectPrice { get; set; }
     
     public virtual ICollection<SemesterSubject> SubjectSemesters { get; set; } = new List<SemesterSubject>();
 }
